tests/wpt/web-platform-tests/browsing-topics/xhr-topics-insecure-context.tentative.http.html

<!doctype html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script>
    promise_test(async t => {
      const xhr = new XMLHttpRequest();

      let xhr_response = new Promise((resolve, reject) => {
        xhr.onreadystatechange = function() {
          if (xhr.readyState == XMLHttpRequest.DONE) {
            // The request was not eligible for topics, as the context is not
            // secure.
            assert_equals(xhr.responseText, "NO_TOPICS_HEADER");
            resolve();
          }
        }
      });

      xhr.open('GET', './resources/check-topics-request-header.py');
      xhr.deprecatedBrowsingTopics = true;
      xhr.send();

      await xhr_response;
    }, 'test XHR in insecure context that sets the deprecatedBrowsingTopics attribtue');
  </script>
</body>
